# prune: speed up using 1<<N instead of math.Pow(2, N)